Refactor: Move MovieDetails and TvShowDetails to UseCase
This commit refactors the codebase by moving the `MovieDetails` and `TvShowDetails` data classes from their respective domain models and remote mappers to their corresponding UseCase files (`GetMovieDetailsUseCase` and `GetTvShowDetailsUseCase`).

Additionally, the following changes were made:

- Simplified `GetTvShowDetailsUseCase` by directly returning the result from `tvShowRepository.getTvShowDetails(tvShowId)`.
- Corrected the `TvShowDetailsRemoteMapper` to use `SeasonRemoteMapper` for mapping seasons.
- Removed an unnecessary query parameter in `TvShowsApiService.getTvShowDetailsById`.
