+++ domain/src/main/java/com/amsterdam/domain/useCase/details/GetMovieDetailsUseCase.kt
@@ -1,8 +1,10 @@
 package com.amsterdam.domain.useCase.details
 
-import com.amsterdam.domain.models.MovieDetails
 import com.amsterdam.domain.repository.MovieRepository
 import com.amsterdam.domain.useCase.common.AddWatchHistoryUseCase
+import com.amsterdam.entity.Actor
+import com.amsterdam.entity.Movie
+import com.amsterdam.entity.Review
 
 class GetMovieDetailsUseCase(
     private val movieRepository: MovieRepository,
@@ -11,4 +13,13 @@ class GetMovieDetailsUseCase(
     suspend operator fun invoke(movieId: Long): MovieDetails {
         return movieRepository.getMovieDetailsById(movieId).also { addWatchHistoryUseCase(movieId) }
     }
+
+    data class MovieDetails(
+        val movie: Movie,
+        val reviews: List<Review>,
+        val actors: List<Actor>,
+        val similarMovies: List<Movie>,
+        val movieGallery: List<String>,
+        val moviePosters: List<String>,
+    )
 }
