+++ remoteDatasource/src/main/java/com/amsterdam/remotedatasource/api/TvShowsApiService.kt
@@ -18,7 +18,7 @@ interface TvShowsApiService {
     @GET(TV_SHOW_DETAILS_ENDPOINT)
     suspend fun getTvShowDetailsById(
         @Path("tvShowId") tvShowId: Long,
-        @Query("append_to_response") appendToResponse: String = TV_SHOW_DETAILS_APPEND_PARAMETERS
+        @Query("append_to_response") appendToResponse: String = TV_SHOW_DETAILS_APPEND_PARAMETERS,
     ): TvShowDetailsRemoteResponse
 
     @GET(TV_SHOW_EPISODES_ENDPOINT)
